unify location computation
