+++ src/platform/packages/shared/kbn-esql-ast/src/__tests__/autocomplete.ts
@@ -20,7 +20,6 @@ import type {
   Location,
   ESQLColumnData,
 } from '../commands_registry/types';
-import { getLocationFromCommandOrOptionName } from '../commands_registry/types';
 import { aggFunctionDefinitions } from '../definitions/generated/aggregation_functions';
 import { groupingFunctionDefinitions } from '../definitions/generated/grouping_functions';
 import { scalarFunctionDefinitions } from '../definitions/generated/scalar_functions';
@@ -210,11 +209,7 @@ export function getFunctionSignaturesByReturnType(
         if (ignoreAsSuggestion) {
           return false;
         }
-        if (
-          !(option ? [...locations, getLocationFromCommandOrOptionName(option)] : locations).some(
-            (loc) => locationsAvailable.includes(loc)
-          )
-        ) {
+        if (!locations.some((loc) => locationsAvailable.includes(loc))) {
           return false;
         }
         const filteredByReturnType = signatures.filter(
