+++ src/platform/packages/shared/kbn-esql-ast/src/commands_registry/types.ts
@@ -151,7 +151,6 @@ export interface ICommandContext {
   histogramBarTarget?: number;
   activeProduct?: PricingProduct | undefined;
 }
-
 /**
  * This is a list of locations within an ES|QL query.
  *
@@ -240,29 +239,3 @@ export enum Location {
    */
   COMPLETION = 'completion',
 }
-
-const commandOptionNameToLocation: Record<string, Location> = {
-  eval: Location.EVAL,
-  where: Location.WHERE,
-  row: Location.ROW,
-  sort: Location.SORT,
-  stats: Location.STATS,
-  inlinestats: Location.STATS,
-  by: Location.STATS_BY,
-  enrich: Location.ENRICH,
-  with: Location.ENRICH_WITH,
-  dissect: Location.DISSECT,
-  rename: Location.RENAME,
-  join: Location.JOIN,
-  show: Location.SHOW,
-  completion: Location.COMPLETION,
-  rerank: Location.RERANK,
-};
-
-/**
- * Pause before using this in new places. Where possible, use the Location enum directly.
- *
- * This is primarily around for backwards compatibility with the old system of command and option names.
- */
-export const getLocationFromCommandOrOptionName = (name: string) =>
-  commandOptionNameToLocation[name];
